Introduce `insertAt()` method
Why is Trilinos so hammer-headed?!?
If a matrix is filled, you can `SumInto` or `Replace`,
but you can't `Insert`.
If a matrix is not filled, you can `Insert` (which will perversely
accumulate values), but you can't `SumInto` or `Replace`.
If you decide that it's silly to have to keep track of all of this
and opt to just create a new matrix every time you need it,
PyTrilinos will [leak at 4 GiB a minute](https://github.com/trilinos/Trilinos/issues/2327).